refactor(vm): split vm_call_dispatch.rs into cohesive modules (§7-8)
Split the 2257-line vm_call_dispatch.rs (over the 500-line limit) into
eight cohesive modules under src/vm/, all under 500 lines. All methods are
impl Interpreter, so the split is visibility-only.

- vm_call_dispatch.rs (294)   - dispatch core: compiled-first, otf compile,
                                interpreter-handled/carrier predicates + tests
- vm_call_resolve.rs (236)    - find_compiled_function + package cache
- vm_call_fast.rs (279)       - call_compiled_function_fast
- vm_call_eligibility.rs (322)- fast/light/share eligibility helpers
- vm_call_light.rs (275)      - positional-light call + type-check helpers
- vm_call_light_typed.rs (361)- call_compiled_function_light + escape check
- vm_call_named.rs (35)       - call_compiled_function_named
- vm_call_named_inner.rs (477)- call_compiled_function_named_inner

A few cross-module private helpers (record_cf_deprecation, is_fast_type_name,
return_value_escapes_routine, type_check_argument_attrs,
call_compiled_function_named_inner) bumped to pub(super). The fn multiset is
byte-identical. No behavior change.
