+++ src/vm.rs
@@ -128,12 +128,19 @@ mod vm_arith_ops;
 mod vm_bitwise_ops;
 mod vm_call_autothread;
 mod vm_call_dispatch;
+mod vm_call_eligibility;
 mod vm_call_exec_ops;
+mod vm_call_fast;
 mod vm_call_func_ops;
 mod vm_call_helpers;
+mod vm_call_light;
+mod vm_call_light_typed;
 mod vm_call_method_compiled;
 mod vm_call_method_mut_ops;
 mod vm_call_method_ops;
+mod vm_call_named;
+mod vm_call_named_inner;
+mod vm_call_resolve;
 mod vm_closure_dispatch;
 mod vm_coerce_concat_ops;
 mod vm_comparison_container_ops;
